requirements.txt
telethon==1.40.0
python-dotenv==1.0.1
aiosqlite==0.19.0
openpyxl==3.1.2